Fix for page titles in wiki page index.
In MySQL, the "group by" clause arbitrarily selects a record to preserve
after aggregation, not necessarily the newest revision of a wiki page.

This could lead to the effect that a change to a wiki page title was not
reflected in the wiki page index, because the wiki page index featured
the title of an arbitrarily decided older revision of the same page.